Refactor data util tests as data handler tests.
